Copied files from hdf5_1_12_1.

+++ tools/test/perform/chunk.c
@@ -6,13 +6,13 @@
  * This file is part of HDF5.  The full HDF5 copyright notice, including     *
  * terms governing use, modification, and redistribution, is contained in    *
  * the COPYING file, which can be found at the root of the source code       *
- * distribution tree, or in https://support.hdfgroup.org/ftp/HDF5/releases.  *
+ * distribution tree, or in https://www.hdfgroup.org/licenses.               *
  * If you do not have access to either file, you may request a copy from     *
  * help@hdfgroup.org.                                                        *
  * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
 /*
- * Programmer:  Robb Matzke <robb@arborea.spizella.com>
+ * Programmer:  Robb Matzke
  *              Thursday, May 14, 1998
  *
  * Purpose:	Checks the effect of various I/O request sizes and raw data
@@ -55,20 +55,20 @@
 #define RM_CACHE_STRT 25
 #define RM_CACHE_END  25
 #define RM_CACHE_DELT 5
-#define RM_START      (double)0.50F
-#define RM_END        (double)5.00F
-#define RM_DELTA      (double)0.50F
-#define RM_W0         0.0F
+#define RM_START      0.50
+#define RM_END        5.00
+#define RM_DELTA      0.50
+#define RM_W0         0.0
 #define RM_NRDCC      521
 
 /* Diagonal test */
 #define DIAG_CACHE_STRT 25
 #define DIAG_CACHE_END  25
 #define DIAG_CACHE_DELT 5
-#define DIAG_START      (double)0.50F
-#define DIAG_END        (double)5.00F
-#define DIAG_DELTA      (double)0.50F
-/* #define DIAG_W0		0.65F */
+#define DIAG_START      0.50
+#define DIAG_END        5.00
+#define DIAG_DELTA      0.50
+/* #define DIAG_W0		0.65 */
 /* #define DIAG_NRDCC		521 */
 
 static size_t nio_g;
